# Conflicts:
#	src/ophys_etl/workflows/app_config/app_config.py
#	src/ophys_etl/workflows/db/initialize_db.py
#	src/ophys_etl/workflows/well_known_file_types.py
#	src/ophys_etl/workflows/workflow_steps.py

+++ src/ophys_etl/modules/sine_dewarp/utils.py
@@ -122,7 +122,7 @@ def xdewarp(imgin: np.ndarray,
     imgout[:, (int(xtable['aL'])):(512-(int(xtable['aR'])))] = \
         imgin[:, (int(xtable['aL'])):(512-(int(xtable['aR'])))]
 
-    col = np.zeros(imgin.shape[0], np.float)
+    col = np.zeros(imgin.shape[0], float)
 
     # Left side
     for j in range(0, int(xtable['aL'])):
